+++ src/main/java/act/aaa/AAAService.java
@@ -24,16 +24,13 @@ import static act.aaa.AAAConfig.ddl;
 import static act.aaa.AAAConfig.loginUrl;
 
 import act.Act;
-import act.app.ActionContext;
-import act.app.App;
-import act.app.AppServiceBase;
+import act.aaa.util.PersistenceServiceCache;
+import act.app.*;
 import act.app.conf.AutoConfig;
 import act.conf.ConfLoader;
 import act.event.OnceEventListenerBase;
 import act.handler.RequestHandler;
-import act.handler.builtin.controller.ActionHandlerInvoker;
-import act.handler.builtin.controller.Handler;
-import act.handler.builtin.controller.RequestHandlerProxy;
+import act.handler.builtin.controller.*;
 import act.handler.builtin.controller.impl.ReflectedHandlerInvoker;
 import act.util.MissingAuthenticationHandler;
 import act.util.SubClassFinder;
@@ -50,9 +47,7 @@ import org.osgl.util.*;
 import org.yaml.snakeyaml.Yaml;
 import osgl.version.Version;
 
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
+import java.io.*;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
 import java.net.URL;
@@ -85,7 +80,7 @@ public class AAAService extends AppServiceBase<AAAService> {
 
     private AuthenticationService authenticationService;
     private AuthorizationService authorizationService;
-    private AAAPersistentService persistentService;
+    private PersistenceServiceCache persistentService;
     private Auditor auditor;
 
     private OnceEventListenerBase onServiceInitialized = new OnceEventListenerBase() {
@@ -244,7 +239,7 @@ public class AAAService extends AppServiceBase<AAAService> {
             // app's implementation should be the winner
             return this;
         }
-        this.persistentService = $.requireNotNull(service);
+        this.persistentService = PersistenceServiceCache.wrap(service);
         if (firstLoadPersistenceService) {
             app().eventBus().trigger(new AAAPersistenceServiceInitialized(this));
         }