+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # act-aaa CHANGE LOG
 
 1.5.2
+* Allow specify mode in the `authenticate.list` #25
 * Support act sample data generation on permission, role and privilege generation #24
 
 1.5.1 - 20/Jun/2018